Separate out implementation of clipboard services
