+++ frontend/src/service/clipboardService/Clipboard.service.ts
@@ -1,0 +1,3 @@
+export interface IClipboardService {
+  copyTextToClipboard(text: string): Promise<void>;
+}
