+++ frontend/src/service/clipboardService/NativeClipboard.service.ts
@@ -1,0 +1,7 @@
+import { IClipboardService } from './Clipboard.service';
+
+export class NativeClipboardService implements IClipboardService {
+  public copyTextToClipboard(text: string): Promise<void> {
+    return navigator.clipboard.writeText(text);
+  }
+}
