+++ frontend/src/service/extensionService/BrowserExtension.service.ts
@@ -1,5 +1,3 @@
-import { EnvService } from '../Env.service';
-
 export interface IBrowserExtensionService {
   isSupported(): boolean;
   isInstalled(): Promise<boolean>;
